requirements.txt
discord.py~=1.4.1
pytz~=2020.1
disputils~=0.1.3
asyncpg~=0.21.0
natural~=0.2.0